requirements.txt
streamlit>=1.28.0
boto3>=1.26.0
pandas>=1.5.0
openai>=1.0.0
sqlparse>=0.4.0
plotly>=5.15.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
